backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
